feat: extend the write suite to DNP; fix two 8-bit truncation bugs
Adds CMD native partitions (DNP) to all four tiers. Doing so exposed two real
engine bugs, both invisible on every floppy format because a CMD native track
holds 256 sectors - exactly one more than a uint8_t can represent, where D64's
largest track is 21.

  #8 initializeBlockAllocationMap() read the sector count into a uint8_t, so
     256 became 0. The bitmap loop then wrote all-zero bytes, marking EVERY
     block allocated on a freshly formatted image, and bitmap_bytes computed
     as 0 which flipped has_count on - making the writer emit a leading count
     byte the reader does not expect, the same writer/reader disagreement as
     finding #6 by another route.

  #9 getTrackFreeCount() returned uint8_t, so a fully free CMD native track
     counted 256 -> 0. blocksFree() reported 0 and, worse, getNextFreeBlock()
     treated every track as full, so no write could allocate anything.

Both now use uint16_t, matching getSectorCount()'s return type. D71's split
side-2 count byte still narrows explicitly, which is correct - those tracks
hold at most 21 sectors.

DNP also needed a directory-location fallback: its constructor reads the
directory T/S from the image at 0x100, which on a zeroed container reads 0/0,
so formatImage() would have written the directory to track 0. It now falls
back to 1/34, where CMD native partitions put it.

FormatFixture gains has_c1541_oracle. VICE cannot attach a DNP, so it is
verified by our seven invariants alone - a genuinely weaker check, made
explicit at every call site rather than letting c1541 fail on an image it
cannot read. Two Tier 2 scenarios skip DNP because they cannot exist there:
directory-full (the disk fills first at any affordable size) and BAM record
boundary (DNP has a single record).

Suite: 21 cases, 21 passing across six formats. Firmware builds clean.

+++ lib/meatloaf/media/disk/d64.cpp
@@ -233,7 +233,11 @@ bool D64MStream::deallocateBlock(uint8_t track, uint8_t sector)
     return setBlockAllocation(track, sector, false);
 }
 
-uint8_t D64MStream::getTrackFreeCount(uint8_t track)
+// Returns uint16_t, not uint8_t: a CMD native track (DNP, DHD) has 256 sectors,
+// so a fully free one counts 256 - which truncates to 0 in a byte. That made
+// blocksFree() report 0 and, worse, made getNextFreeBlock() treat every track as
+// full so no write could allocate anything.
+uint16_t D64MStream::getTrackFreeCount(uint8_t track)
 {
     BAMRecord rec;
     uint8_t buf[32];
@@ -244,7 +248,7 @@ uint8_t D64MStream::getTrackFreeCount(uint8_t track)
         return buf[0];
 
     // Bitmap-only record: count the free (1) bits
-    uint8_t count = 0;
+    uint16_t count = 0;
     for (uint8_t i = 0; i < rec.byte_count; i++)
         count += std::bitset<8>(buf[i]).count();
     return count;
